+++ db/upgrade.php
@@ -5,6 +5,14 @@
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 
 /**
  * Upgrade steps for the Flip page activity.
@@ -23,5 +31,9 @@ defined('MOODLE_INTERNAL') || die();
  * @return bool
  */
 function xmldb_flippage_upgrade($oldversion) {
+    if ($oldversion < 2026073000) {
+        upgrade_mod_savepoint(true, 2026073000, 'flippage');
+    }
+
     return true;
 }